Refactor the auth bits into something that can be used in multiple places
